refactor: added psr12 formating
