+++ test/unit/model/DataStore/DeliveryMetadataListenerTest.php
@@ -1,8 +1,24 @@
 <?php
 
+/**
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; under version 2
+ * of the License (non-upgradable).
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ *
+ * Copyright (c) 2021  (original work) Open Assessment Technologies SA;
+ */
 
-namespace oat\taoDeliveryRdf\test\unit\model\assembly;
-
+namespace oat\taoDeliveryRdf\test\unit\model\DataStore;
 
 use core_kernel_persistence_smoothsql_SmoothModel;
 use oat\generis\test\OntologyMockTrait;
@@ -33,11 +49,11 @@ class DeliveryMetadataListenerTest extends TestCase
     /** @var core_kernel_persistence_smoothsql_SmoothModel */
     private $ontology;
 
-    /** @var DeliveryMetadataListener */
-    private $subject;
-
     /** @var FeatureFlagChecker|MockObject */
     private $log;
+
+    /** @var DeliveryMetadataListener */
+    private $subject;
 
     protected function setUp(): void
     {
@@ -63,7 +79,7 @@ class DeliveryMetadataListenerTest extends TestCase
     {
 
         $class = $this->ontology->getClass('http://tao.tld/bogusUri');
-        $mockDelivery = $class->createInstance('Bogus');
+        $class->createInstance('Bogus');
 
         $event = $this->createMock(DeliveryCreatedEvent::class);
         $event->expects($this->once())->method('getDeliveryUri');
