+++ test/unit/model/DataStore/MetaDataDeliverySyncTaskTest.php
@@ -1,8 +1,24 @@
 <?php
 
+/**
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; under version 2
+ * of the License (non-upgradable).
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ *
+ * Copyright (c) 2021  (original work) Open Assessment Technologies SA;
+ */
 
-namespace oat\taoDeliveryRdf\test\unit\model\assembly;
-
+namespace oat\taoDeliveryRdf\test\unit\model\DataStore;
 
 use core_kernel_persistence_smoothsql_SmoothModel;
 use oat\generis\model\data\Ontology;
@@ -59,7 +75,7 @@ class MetaDataDeliverySyncTaskTest extends TestCase
             Ontology::SERVICE_ID => $this->ontology
         ]);
 
-        $this->subject = new MetaDataDeliverySyncTask;
+        $this->subject = new MetaDataDeliverySyncTask();
     }
 
     public function testJsonSerialize(): void
@@ -70,7 +86,7 @@ class MetaDataDeliverySyncTaskTest extends TestCase
         );
     }
 
-    public function test__invoke()
+    public function testInvoke()
     {
         $this->persistDataService->method('persist');
         $this->qtiTestService->method('getItems')->willReturn([]);
